requirements.txt
Flask==2.2.5
web3==6.10.0
py-solc-x==1.1.0